Resolve fixme
