--- hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/config/SharedConfigDstu3Plus.java
@@ -1,56 +1,0 @@
-package ca.uhn.fhir.jpa.config;
-
-/*-
- * #%L
- * HAPI FHIR JPA Server
- * %%
- * Copyright (C) 2014 - 2022 Smile CDR, Inc.
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-
-import ca.uhn.fhir.jpa.dao.ObservationLastNIndexPersistSvc;
-import ca.uhn.fhir.jpa.term.TermCodeSystemStorageSvcImpl;
-import ca.uhn.fhir.jpa.term.TermConceptDaoSvc;
-import ca.uhn.fhir.jpa.term.TermDeferredStorageSvcImpl;
-import ca.uhn.fhir.jpa.term.TermReindexingSvcImpl;
-import ca.uhn.fhir.jpa.term.api.ITermCodeSystemStorageSvc;
-import ca.uhn.fhir.jpa.term.api.ITermDeferredStorageSvc;
-import ca.uhn.fhir.jpa.term.api.ITermReindexingSvc;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-// FIXME: rename since this is DSTU2 plus
-@Configuration
-public class SharedConfigDstu3Plus {
-
-	@Bean
-	public ITermCodeSystemStorageSvc termCodeSystemStorageSvc() {
-		return new TermCodeSystemStorageSvcImpl();
-	}
-
-
-
-	@Bean
-	public ITermReindexingSvc termReindexingSvc() {
-		return new TermReindexingSvcImpl();
-	}
-
-	@Bean
-	public ObservationLastNIndexPersistSvc baseObservationLastNIndexpersistSvc() {
-		return new ObservationLastNIndexPersistSvc();
-	}
-
-
-}
