+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/config/r4b/JpaR4BConfig.java
@@ -6,7 +6,6 @@ import ca.uhn.fhir.jpa.api.IDaoRegistry;
 import ca.uhn.fhir.jpa.api.dao.IFhirSystemDao;
 import ca.uhn.fhir.jpa.config.GeneratedDaoAndResourceProviderConfigR4B;
 import ca.uhn.fhir.jpa.config.JpaConfig;
-import ca.uhn.fhir.jpa.config.SharedConfigDstu3Plus;
 import ca.uhn.fhir.jpa.dao.ITransactionProcessorVersionAdapter;
 import ca.uhn.fhir.jpa.dao.r4b.TransactionProcessorVersionAdapterR4B;
 import ca.uhn.fhir.jpa.graphql.GraphQLProvider;
@@ -53,7 +52,6 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 @Import({
 	FhirContextR4BConfig.class,
 	GeneratedDaoAndResourceProviderConfigR4B.class,
-	SharedConfigDstu3Plus.class,
 	JpaConfig.class
 })
 public class JpaR4BConfig {
